Improvements to implicit conversion handling in Organize Imports.
Implicit conversion, e.g. like the JavaConversions are now detected
(more) correctly and not seen as unused. I fear that there are quite
a few errors lurking in the implicits corner, but to me it seems quite
complicated to determine if an import for an implicit is needed or not.