Some refactoring and memoization to fight the recent performance losses.
This might help counter the recently introduced performance problems,
a superficial measurement shows that running the tests on my machine
is a bit faster.

As a next step, it might make sense to convert the vanilla Scala AST
to a pimped refactoring AST once at the beginning of the refactoring
instead of doing the same work again and again in extractors.