Scala 2.10/2.11 cross compatibility for a test
In Scala 2.11, the presentation compiler will no longer
see macro expansions in trees. Instead, it will see the
expandee (e.g. `StringContext("").f()`, attributed with
the type of the expansion.

The output of one test in `CompilationUnitDependenciesTest`
will change as result of this.

This commit changes that test so that it no longer includes
a macro call (the call to the class tag materializer). This
means the test will work in Scala 2.10 and 2.11.