Handle strings properly when stipping comments
The problem described in ticket #1002166 comes from the fact that
`CommentsUtils.stripComment` incorrectly treats `//` in the string constant "//"
as single line comment. Therefore `("//")` is transformed to `("`. This faulty
result is then used by `CommonPrintUtils.balanceParens`, where it leads to
surplus `)` being inserted. This commit makes sure that this doesn't happen
any more.

Fixes #1002166