Remove code that seems to have no functionality
`CommonPrintUtils.balanceBrackets` used to contain some code that was meant to
add missing closing brackets directly after the last closing bracket, instead of
just appending them to the layout. Given the fact that the code was broken anyway,
(it didn't consider comments, string constants, etc.), all tests pass without it,
and the original author thinks that it's not needed, it seems safe to remove it.