Revert "Remove Eclipse update-site and feature"
This reverts commit dd8caa2ecdcce8de39e55f0c86c56e2d59461fcd.

As it turned out, I was wrong. The IDE very well depends on the
update-site. Its maven build needs the update site to find the
scala-refactoring dependency. The better fix is to not stress
scala-refactoring with the update-site but to fix the Scala IDE build.
Until this is done, the change is reverted in order to not block the
current IDE builds.