Minor, but subtle change to `Movements.until`
With this commit, the movement `until` is meant to search for, is tried
*before* eventually skipping parts of the input. Prior to this commit, it
was the other way round. Take a look at the ScalaDocs that come with
this commit, if you have trouble understanding what this means.

Note that I did not find an example where the old behaviour led to a bug
in a refactoring, but I expect that there might be cases involving backtick
identifiers, where this might matter.