Don't drop return types on rename refactorings
Prior versions of the library contained logic that compared return types from
ASTs to the corresponding strings in the source code to decide if these types
should be printed or not. This didn't work properly because a simple sting
comparison was used, that could be easily mislead (see ticket #1002560)
since a single type might have many different equivalent string representations
in the source like `Map[Int,Int]`, `Map[Int, Int]`, `Map[ Int, Int ]` or even
`Map[/*Long*/Int]`. Another example would be `Seq[Bug]` and `Seq[test.Bug]`, given that
`Bug` is from package `test`. This commit drops most of this logic, since for
legal Scala code, we can assume that a type is present in the source code, if
it has a range position. Apart from that, we also handle the case where a
method annotated with `override` is missing a return type.

Fixes #1002560