Correctly set positions for selects involving package objects
For whatever reason, qualifiers from `Select` trees involving package objects
have the same `RangePosition` as the `Select` trees themselves. When renaming
a member of a package object, the generated `NameTrees` with the new name
got the wrong `Position`, which resulted in them not being correctly associated
with the original `NameTrees`. This is the root cause for ticket #1002609.

This commit fixes the aforementioned problem by correctly setting these
positions. Interestingly this makes the code simpler as well.

Fixes #1002609