Be more careful when setting positions
When setting positions for `NameTrees` that are tied to `Selects`, the
the assumption that the name starts where the `Select` tree starts might be
wrong, for example if default arguments and comments are involved, as in
ticket #1002611. The only reliable way to set positions in this case is to
extract them from the original source code, which is exactly what this
commit does.