Remove cod coverage plugin
scala-refactoring needs to build against nightlies of scalac, for which
scoverage unfortunately is not available. We don't need to run the
plugin on the Scala PR CI but unfortunately it is still resolved. We
first need to find out how to disable the resolution at all before we
can re-enable that plugin.