Organize Imports in Defs avoiding tree printer
Attempt to organize imports with no modification of original tree.
There are 2 files which are the machinery:
1 DefImportsOrganizer.scala
2 NotPackageImportPratitioners.scala
The first one is the heart of engine

An example of usage is in OrganizeImports.scala. Main idea concerns
creation of `TextChanges` and adding them to those produced by
"PackageDef" old fashionable part.