Remove unused imports in source folder
All changes were done automatically by the following script:

    tac warnings.txt | awk '{split($0,a,":"); print "'"'"'sed -i " "\""a[2] "d\" " a[1] "'"'"'"}' | xargs -i -t sh -c "eval {}"

`warnings.txt` is a file whose contents look like this:

    /path-to-refactoring/scala-refactoring/src/main/scala/scala/tools/refactoring/Refactoring.scala:10: Unused import
    /path-to-refactoring/scala-refactoring/src/main/scala/scala/tools/refactoring/common/EnrichedTrees.scala:11: Unused import
    /path-to-refactoring/scala-refactoring/src/main/scala/scala/tools/refactoring/common/Selections.scala:9: Unused import
    /path-to-refactoring/scala-refactoring/src/main/scala/scala/tools/refactoring/common/Selections.scala:11: Unused import
    /path-to-refactoring/scala-refactoring/src/main/scala/scala/tools/refactoring/common/TreeTraverser.scala:8: Unused import

These are the actual error messages of the compiler or better the first
line of each error message, since this line contains all the important
information. The script generates a command that calls `sed` to remove
the import.